cleanup wallet
